project/backend/requirements.txt
python-dotenv
langchain
langchain-core
langchain-tavily
langchain_google_genai
pydantic
httpx
trafilatura
asyncio